Refactor process service by pulling the raw process APIs out of the class
